+++ src/client/index.ts
@@ -1,11 +1,9 @@
 import { Game } from "@client/game";
 import { View } from "@client/view";
-import io from "socket.io-client";
 
 const DEBUG_MODE = false;
 
-// Create socket.io connection to server
-const socket = io();
+
 
 // Create view
 const view = new View();