+++ src/share/action/Action.ts
@@ -1,0 +1,8 @@
+/**
+ * An action is any change to the state of the game.
+ */
+export abstract class Action {
+  constructor() {
+
+  }
+}