+++ src/share/action/block/RemoveBlockAction.ts
@@ -1,0 +1,10 @@
+import { IAbsoluteCoordinate } from "@share/data/coordinate";
+import { Action } from "../Action";
+
+export class RemoveBlockAction extends Action {
+  constructor(
+    coord: IAbsoluteCoordinate
+  ) {
+    super();
+  }
+}