+++ src/share/action/cluster/LoadClusterAction.ts
@@ -1,0 +1,10 @@
+import { IClusterData } from "@share/data/cluster-data";
+import { Action } from "../Action";
+
+export class LoadClusterAction extends Action {
+  constructor(
+    cluster: IClusterData
+  ) {
+    super();
+  }
+}