+++ src/share/action/cluster/UnloadClusterAction.ts
@@ -1,0 +1,7 @@
+import { Action } from "../Action";
+
+export class UnloadClusterAction extends Action {
+  constructor() {
+    super();
+  }
+}