+++ changelog.md
@@ -3,6 +3,7 @@
 * Fix player movement behavior (going under arches, placing blocks under self, etc.)
 * Reuse code between coordinate classes
 * Remove error throwing from stringify methods (return null on fail)
+* Move IPlayerMotor inside of IPlayer system
 
 ## 0.1.1